requirements.txt
scipy>=0.13.3
scikit-learn==0.19.1
pandas>=0.23
numpy>=1.8.2
cognite-model-hosting
cognite-sdk
matplotlib
jupyter
tornado<6.0
